+++ Sources/Moya/TargetType.swift
@@ -28,7 +28,5 @@ public protocol TargetType {
 public extension TargetType {
 
     /// The type of validation to perform on the request. Default is `.none`.
-    var validationType: ValidationType {
-        return .none
-    }
+    var validationType: ValidationType { .none }
 }