+++ Tests/MoyaTests/MoyaProviderIntegrationTests.swift
@@ -11,7 +11,7 @@ import OHHTTPStubsSwift
 @testable import ReactiveMoya
 
 func beIdenticalToResponse(_ expectedValue: Moya.Response) -> Predicate<Moya.Response> {
-    return Predicate { expression in
+    Predicate { expression in
         let test: Bool
         if let value = try expression.evaluate(), value == expectedValue {
             test = true
@@ -119,18 +119,19 @@ final class MoyaProviderIntegrationTests: QuickSpec {
                     }
 
                     it("uses a background queue") {
-                        var isMainThread: Bool?
+
+                        let callbackQueueLabel = Atomic<String?>(wrappedValue: nil)
                         let callbackQueue = DispatchQueue(label: "background_queue", attributes: .concurrent)
                         let target: GitHub = .zen
 
                         waitUntil { done in
                             provider.request(target, callbackQueue: callbackQueue) { _ in
-                                isMainThread = Thread.isMainThread
+                                callbackQueueLabel.wrappedValue = DispatchQueue.currentLabel
                                 done()
                             }
                         }
 
-                        expect(isMainThread) == false
+                        expect(callbackQueueLabel.wrappedValue) == "background_queue"
                     }
 
                     it("uses the main queue") {