+++ Changelog.md
@@ -1,5 +1,7 @@
 # Next
 
+### Changed
+- Moya's Swift version is now Swift 5.1. [#1931](https://github.com/Moya/Moya/pull/1931) by [@BasThomas](https://github.com/BasThomas).
 # [14.0.0] - 2020-02-15
 
 ### Changed